+++ out/Core/CLANGx64.Blitz.Console/build_info.h
@@ -2,8 +2,8 @@
 #define bmMONTH  6
 #define bmDAY    19
 #define bmHOUR   13
-#define bmMINUTE 38
-#define bmSECOND 13
-#define bmTIME   Time(2026, 6, 19, 13, 38, 13)
+#define bmMINUTE 59
+#define bmSECOND 57
+#define bmTIME   Time(2026, 6, 19, 13, 59, 57)
 #define bmMACHINE "DESKTOP-GFX1"
 #define bmUSER    "CodexSandboxOffline"