+++ out/tests/StateMachineCoreTest/CLANGx64.Blitz.Console.Main/build_info.h
@@ -2,11 +2,11 @@
 #define bmMONTH  6
 #define bmDAY    19
 #define bmHOUR   13
-#define bmMINUTE 38
-#define bmSECOND 14
-#define bmTIME   Time(2026, 6, 19, 13, 38, 14)
+#define bmMINUTE 59
+#define bmSECOND 57
+#define bmTIME   Time(2026, 6, 19, 13, 59, 57)
 #define bmMACHINE "DESKTOP-GFX1"
 #define bmUSER    "CodexSandboxOffline"
-#define bmGIT_REVCOUNT "25"
-#define bmGIT_HASH "298742bab1a3d4b76d4c770134f6392a4fc74b41"
+#define bmGIT_REVCOUNT "26"
+#define bmGIT_HASH "81705c1184e5a204527d1374efed1c82d7954d0f"
 #define bmGIT_BRANCH "main"